chore: move modal store into global stores folder
